Fix segfault on overload resolution
In some special cases, the overload
resolution with the overload declarations
may cause some diagnostics to be
falsly ignored. This will lead to
a code that we assume as correct, but
actually it's not. Fix the overload
resolution, by correcting the flaws.

Issue: https://gitee.com/openharmony/arkcompiler_runtime_core/issues/ICWZWE

Internal issue: #29382

Test: build

Change-Id: Ib136a5ee4748edcbdb1735e94c2250d6a4ecaf46
Signed-off-by: Martin Sajti <martin.sajti@h-lab.eu>