Fix readonly field inheritance generic interface
When a generic class inherits a generic interface
with readonly field declared, the compiler generates
the corresponding getter/setter method for them.
The setter must be generated to allow the assignment
in the constructor. If the type of the field is
the generic type parameter, and it has an 'out'
variance modifier, then CTE is thrown, as setter
tries to use the type parameter in the in position
for the setter.
The patch skips the variance check for these
generated setters, as these are generated to
be safe, and invisible to the user.

Issue: https://gitcode.com/openharmony/arkcompiler_ets_frontend/issues/8069

Internal issue: #24444

Test: build, cts test

Change-Id: Ic91f7df4081deaa1f294f3b0f4ac5cfb3e4692c9
Signed-off-by: Martin Sajti <martin.sajti@h-lab.eu>